refactor: error and imports
