+++ src/commands/database/database-reset.persistence.ts
@@ -1,7 +1,7 @@
 import { TaskEither, tryCatch as taskEitherTryCatch } from 'fp-ts/TaskEither';
 import type { PoolClient, QueryResult } from 'pg';
 import type { PostgresDb } from '@fastify/postgres';
-import { Errors, InfrastructureError } from '../../reporter/http-reporter';
+import { Errors, InfrastructureError } from '../../reporter';
 
 export const resetDatabaseStructure = (database: PostgresDb): TaskEither<Errors, QueryResult> =>
   taskEitherTryCatch(dropAndRecreateTables(database)(), onDropAndRecreateTablesError);
