+++ src/commands/edit-fare/edit-fare.ts
@@ -1,6 +1,6 @@
 import { pipe } from 'fp-ts/lib/function';
 import { map as taskEitherMap, TaskEither } from 'fp-ts/TaskEither';
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 import { EditedToPersist, FaresToEdit } from './edit-fare.route';
 import { isOneWay } from '../../domain/utils';
 import { toPending } from '../../mappers';
