+++ src/commands/schedule-fare/schedule-fare.ts
@@ -3,7 +3,7 @@ import { Either, map as eitherMap } from 'fp-ts/Either';
 import { FaresToSchedulePersist, FareToSchedule } from './schedule-fare.route';
 import { isOneWay } from '../../domain/utils';
 import { toPending } from '../../mappers';
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 
 export const scheduleFare = (fareToSchedule: Either<Errors, FareToSchedule>): Either<Errors, FaresToSchedulePersist> =>
   pipe(fareToSchedule, eitherMap(applySchedule));
