+++ src/commands/schedule-pending/schedule-pending.persistence.ts
@@ -1,11 +1,11 @@
 import { map as taskEitherMap, TaskEither, tryCatch as taskEitherTryCatch } from 'fp-ts/TaskEither';
 import type { PoolClient, QueryResult } from 'pg';
 import type { PostgresDb } from '@fastify/postgres';
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 import { Entity, Scheduled } from '../../definitions';
 import { PendingToSchedulePersist } from './schedule-pending.route';
 import { pipe } from 'fp-ts/function';
-import { onDatabaseError } from '../../reporter/database.error';
+import { onDatabaseError } from '../../errors';
 import { fromDBtoPendingCandidate, fromDBtoScheduledCandidate } from '../../persistence/persistence-utils';
 
 export const persistPendingScheduled =
