+++ src/commands/subcontract-fare/subcontract-fare.validation.ts
@@ -1,4 +1,4 @@
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 import { pipe } from 'fp-ts/lib/function';
 import { chain as taskEitherChain, fromEither, TaskEither, tryCatch as taskEitherTryCatch } from 'fp-ts/TaskEither';
 import { PostgresDb } from '@fastify/postgres';
@@ -15,8 +15,8 @@ import {
 } from '../../codecs';
 import { type as ioType, Type, union as ioUnion } from 'io-ts';
 import { FaresSubcontracted, FaresToSubcontract } from './subcontract-fare.route';
-import { throwEntityNotFoundValidationError } from '../../reporter/entity-not-found.validation-error';
-import { $onInfrastructureOrValidationError } from '../../reporter/infrastructure-or-validation.error';
+import { throwEntityNotFoundValidationError } from '../../errors/entity-not-found.validation-error';
+import { $onInfrastructureOrValidationError } from '../../errors/infrastructure-or-validation.error';
 import { fromDBtoScheduledCandidate } from '../../persistence/persistence-utils';
 
 export const $subcontractFareValidation =
