+++ src/queries/estimate-journey/estimate-journey.validation.ts
@@ -2,7 +2,7 @@ import { pipe } from 'fp-ts/function';
 import { JourneyTransfer, journeyTransferCodec } from './estimate-journey.codec';
 import { TaskEither } from 'fp-ts/lib/TaskEither';
 import { fromEither } from 'fp-ts/TaskEither';
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 import { externalTypeCheckFor, journeyCodec, journeyRulesCodec } from '../../codecs';
 import { Journey } from '../../definitions';
 import { chain as eitherChain, Either } from 'fp-ts/Either';
