+++ src/queries/predict-recurrence/predict-recurrence.ts
@@ -1,5 +1,5 @@
 import { TaskEither } from 'fp-ts/TaskEither';
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 import { PredictedRecurrence, PredictRecurrence } from '../../definitions';
 
 export type PredictRecurrenceAdapter = (predict: PredictRecurrence) => TaskEither<Errors, PredictedRecurrence>;
