+++ src/queries/predict-recurrence/predict-recurrence.validation.ts
@@ -1,7 +1,7 @@
 import { pipe } from 'fp-ts/function';
 import { TaskEither } from 'fp-ts/lib/TaskEither';
 import { fromEither } from 'fp-ts/TaskEither';
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 import { PredictRecurrence } from '../../definitions';
 import { predictRecurrenceCodec } from '../../codecs';
