+++ src/queries/search-place/search-place.validation.ts
@@ -1,7 +1,7 @@
 import { pipe } from 'fp-ts/function';
 import { TaskEither } from 'fp-ts/lib/TaskEither';
 import { fromEither } from 'fp-ts/TaskEither';
-import { Errors } from '../../reporter/http-reporter';
+import { Errors } from '../../reporter';
 import { stringCodec } from '../../codecs';
 
 export const searchPlaceValidation = (transfer: unknown): TaskEither<Errors, string> =>
